+++ physicalai/__init__.py
@@ -9,6 +9,7 @@ from physicalai.envs.tabletop import TabletopEnv, parse_instruction, ACTIONS, AC
 from physicalai.envs.gridworld import GridWorld, GW_ACTIONS
 from physicalai.policies.base import Policy, ScriptedPolicy, RandomPolicy
 from physicalai.policies.learned import BCPolicy, collect_expert_demos, featurize
+from physicalai.policies.vlm import VLMPolicy, describe
 from physicalai.eval.harness import Scorecard, evaluate, evaluate_split
 from physicalai.rl.tabular import train_q, GreedyQPolicy
 
@@ -17,6 +18,7 @@ __all__ = [
     "GridWorld", "GW_ACTIONS",
     "Policy", "ScriptedPolicy", "RandomPolicy",
     "BCPolicy", "collect_expert_demos", "featurize",
+    "VLMPolicy", "describe",
     "Scorecard", "evaluate", "evaluate_split",
     "train_q", "GreedyQPolicy",
 ]